feat: implement authentication context and hooks, refactor imports, and enhance user profile fetching
