+++ src/auth/AuthContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const AuthContext = createContext();
+
+export default AuthContext;
