+++ src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.jsx";
 import "./styles/tailwind.css";
-import { AuthProvider } from "./auth/useAuth";
+import AuthProvider from "./auth/AuthProvider.jsx";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
