+++ src/pages/Homepage.jsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { Link } from "react-router-dom";
 
 export default function Homepage() {
